requirements.txt
PyGithub==1.55
transformers==4.30.0
python-dotenv==0.19.0
smtplib
python-dotenv